+++ cuda_functions/nms_3D/src/cuda/nms_kernel.h
@@ -6,7 +6,7 @@ extern "C" {
 #endif
 
 #define DIVUP(m,n) ((m) / (n) + ((m) % (n) > 0))
-int const threadsPerBlock = sizeof(unsigned long long) * 8;
+int const threadsPerBlock = sizeof(int64_t) * 8;
 
 void _nms(int boxes_num, float * boxes_dev,
           int64_t * mask_dev, float nms_overlap_thresh);
